+++ api/v1/serviceaccount_request_types.go
@@ -12,15 +12,14 @@ const (
 )
 
 const (
-	ConditionTypeProducerReady                   = "ProducerReady"
-	ConditionReasonProducerReadyProducerFound    = "ProducerFound"    // producer found
-	ConditionReasonProducerReadyProducerNotFound = "ProducerNotFound" // producer not found
-	ConditionReasonProducerReadyProducerNotReady = "ProducerNotReady" // producer is found but not ready
-	ConditionTypeProducer                        = "Accepted"
-	ConditionReasonAcceptedValidationSuccess     = "ValidationSuccess"
-	ConditionReasonAcceptedInvalidParameters     = "InvalidParameters"
-	ConditionReasonReadyServiceAccountCreated    = "ServiceAccountCreated"
-	ConditionReasonReadyServiceAccountFailed     = "ServiceAccountFailed" // error during http call or exec
+	ConditionTypeProducerReady                          = "ProducerReady"
+	ConditionReasonProducerReadyProducerFound           = "ProducerFound"    // producer found
+	ConditionReasonProducerReadyProducerNotFound        = "ProducerNotFound" // producer not found
+	ConditionReasonProducerReadyProducerNotReady        = "ProducerNotReady" // producer is found but not ready
+	ConditionTypeServiceAccountReady                    = "ServiceAccountReady"
+	ConditionReasonServiceAccountReadyCreated           = "ServiceAccountCreated" // the service-account is created
+	ConditionReasonServiceAccountReadyFailed            = "ServiceAccountFailed"  // error during http call or exec
+	ConditionReasonServiceAccountReadyInvalidParameters = "InvalidParameters"     // spec validation failed
 )
 
 // LocalSecretRef definiert eine Referenz auf ein Secret im selben Namespace.