+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- [#3] Improve status conditions for ServiceAccountRequests
+
 ## [v1.0.0] - 2026-04-22
 ### Added
 - [#1] Initial CRDs for ServiceAccountRequest and ServiceAccountProducer