+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 - [#3] Improve status conditions for ServiceAccountRequests
-- [#3] Simplify serviceaccount-request params to just an array of strings
+- [#3] Simplify serviceaccount-request params to just a map of strings
+  - When creating service accounts, the params are passed to the serviceaccount-producer via the serviceaccount operator
 
 ## [v1.0.0] - 2026-04-22
 ### Added